New unit test: test_analytic_init.py
