+++ tests/main/driver/test_example_configs.py
@@ -19,7 +19,7 @@ TESTED_CONFIGS: List[str] = [
     "baroclinic_c12_null_comm.yaml",
     "baroclinic_c12_write_restart.yaml",
     "baroclinic_c48_6ranks_serialbox_test.yaml",
-    "analytic_test.yaml"
+    "analytic_test.yaml",
 ]
 EXCLUDED_CONFIGS: List[str] = [
     # We don't test serialbox example because it loads namelist
